+++ Assets/Scripts/Items/ItemNumOne.cs
@@ -8,7 +8,7 @@ public class ItemNumOne : Item
     [SerializeField][Range(0, 10)] private int maxLevel;
     [SerializeField][Range(0, 10)] protected int level;
 
-    [SerializeField] private int baseCost;
+    [SerializeField] private long baseCost;
     [SerializeField] protected long cost;
     [SerializeField] protected long itemValue;
